+++ src/functions/quota-collector/main.py
@@ -12,6 +12,7 @@ from modules.qmchecks.ec2.ec2 import get_current_quotastatus_ec2
 from modules.qmchecks.ebs import get_current_quotastatus_ebs
 from modules.qmchecks.ec2_ipam import get_current_quotastatus_ec2_ipam
 from modules.qmchecks.customer_profiles import get_current_quotastatus_customer_profiles
+from modules.qmchecks.chime import get_current_quotastatus_chime
 from modules.qmchecks.drs import get_current_quotastatus_drs
 from modules.qmchecks.schemas import get_current_quotastatus_schemas
 from modules.qmchecks.launchwizard import get_current_quotastatus_launchwizard
@@ -230,6 +231,7 @@ def lambda_handler(event, context):
     entries.extend(get_current_quotastatus_ebs(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_ec2_ipam(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_customer_profiles(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_chime(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_drs(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_schemas(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_launchwizard(ctx=ctx, skip=official))