+++ src/modules/qmcore/registry.py
@@ -7,6 +7,7 @@ _CHECK_MODULES = (
     ('modules.qmchecks.ebs', 'ebs'),
     ('modules.qmchecks.ec2_ipam', 'ec2-ipam'),
     ('modules.qmchecks.customer_profiles', 'profile'),
+    ('modules.qmchecks.chime', 'chime'),
     ('modules.qmchecks.drs', 'drs'),
     ('modules.qmchecks.schemas', 'schemas'),
     ('modules.qmchecks.launchwizard', 'launchwizard'),